Extract SpecClassDiscovery
